move oauthserver scope package to point of use
