+++ hack/import-restrictions.json
@@ -423,8 +423,6 @@
       "vendor/k8s.io/klog",
       "vendor/github.com/davecgh/go-spew/spew",
       "github.com/openshift/origin/pkg/authorization/authorizer/scopelibrary",
-      "github.com/openshift/origin/pkg/oauth/apis/oauth/validation",
-      "github.com/openshift/origin/pkg/oauth/scope",
       "github.com/openshift/origin/pkg/cmd/server/apis/config",
       "github.com/openshift/origin/pkg/cmd/server/apis/config/install",
       "github.com/openshift/origin/pkg/cmd/server/apis/config/latest",
