+++ pkg/oauthserver/scopecovers/scope.go
@@ -1,4 +1,4 @@
-package scope
+package scopecovers
 
 import (
 	"sort"
