+++ pkg/oauthserver/scopecovers/scope_test.go
@@ -1,4 +1,4 @@
-package scope
+package scopecovers
 
 import (
 	"reflect"
